+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-### 0.0.1 017-02-17
-- Start git bash in vsc
+### 0.0.5 2017-5-18
+- add start shell feature, so `startgit` feature can be replaced
+- change `startgit` configuration from `startgit.shellPath` to `startgit.path`
 
 ### 0.0.2 017-02-17
 - Add icon
+
+### 0.0.1 017-02-17
+- Start git bash in vsc